refactor(vm): extract CLI into ironplc-vm-cli crate
Move the CLI binary, logger, and integration tests from ironplc-vm into
a new ironplc-vm-cli crate. The VM library crate now contains only the
core VM engine with no CLI dependencies (clap, ctrlc, env_logger, time).

This separation is Phase 2 of the no_std VM plan — the VM crate's
dependency surface shrinks, preparing it for the no_std transition in
later phases.
