+++ compiler/vm-cli/resources/test/steel_thread.iplc
