+++ compiler/vm-cli/src/logger.rs
