+++ compiler/vm/Cargo.toml
@@ -9,18 +9,3 @@ repository.workspace = true
 
 [dependencies]
 ironplc-container = { path = "../container", version = "0.158.0" }
-ctrlc = "3"
-
-time = "0.3.17"
-clap = { version = "4.0", features = ["derive", "wrap_help"] }
-env_logger = "0.10.0"
-log = "0.4.20"
-
-[dev-dependencies]
-assert_cmd = { version = "2.0" }
-predicates = { version = "3.0" }
-tempfile = "3"
-
-[[bin]]
-name = "ironplcvm"
-path = "bin/main.rs"
