+++ compiler/vm/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod cli;
 pub mod error;
-pub mod logger;
 pub(crate) mod scheduler;
 pub(crate) mod stack;
 pub(crate) mod value;
